Resolve S11P12D210-94 "Lecture upload"

See merge request s11-webmobile4-sub2/S11P12D210!72

+++ app/src/main/java/com/ssafy/yoganavi/ui/homeUI/myPage/managementVideo/lecture/LectureViewHolder.kt
@@ -9,7 +9,7 @@ import com.ssafy.yoganavi.ui.utils.toK
 
 class LectureViewHolder(
     private val binding: ListItemLectureThumbnailBinding,
-    private val navigateToRegisterVideoFragment: (Int) -> Unit
+    private val navigateToRegisterVideoFragment: (Long) -> Unit
 ) : ViewHolder(binding.root) {
 
     private var likeCount = 0
